requirements.txt
Pillow
PyPDF2
python-docx
openpyxl
